feat(play): control frame rate
The expected frame rate is 30 fps, but since it is hard to measure and
control the time of reading the first line of the frame with BufReader,
the frame rate is set to 32 fps (a little bit faster than 30 fps). The
video will be slightly faster than the audio, so an deceleration in
Final Cut Pro is needed.